split xptokendetail screen with nonxptoken with scrollscreen
