+++ packages/core-mobile/app/new/features/activity/components/ActivityList.tsx
@@ -99,7 +99,7 @@ export const ActivityList = ({
   )
 }
 
-const SectionHeader = ({
+export const SectionHeader = ({
   title,
   index
 }: {
